Assigned datatype to outermost SELECT statement in models.
* 3 models updated

* Assigned datatype to outermost SELECT statement in models.

+++ stroma/models/argentum/silver__episode_event.sql
@@ -11,7 +11,7 @@ MODEL (
   ));
 
 SELECT
-  ee.episode_id,
-  ee.event_id,
-  ee.episode_event_field_concept_id
+  ee.episode_id::INT,
+  ee.event_id::INT,
+  ee.episode_event_field_concept_id::INT
 FROM bronze.episode_event AS ee
